[#15] creating group to organize commands and helping users with a clean and documented interface
